requirements.txt
websocket-client==1.7.0
websocket-server==0.6.4
matplotlib==3.7.2
loguru==0.7.0
protobuf==4.25.2
carla==0.9.15
transforms3d==0.3.1
numpy==1.20.0
sympy==1.12
